requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
h11==0.14.0
packaging==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
uvicorn==0.30.0
waitress==3.0.0
whitenoise==6.6.0
